+++ lib/generators/mokio/templates/mokio.rb
@@ -139,6 +139,7 @@ Mokio.setup do |config|
   #
   # config.frontend_facebook_app_id = ''
 
+<<<<<<< HEAD
 
 
   # External parameters for mokio controllers
@@ -151,5 +152,25 @@ Mokio.setup do |config|
 
   # mattr_accessor :mokio_gems_parameters
   # self.mokio_gems_parameters = {}
+=======
+  #
+  # You can set where the user is to be redirected after logging in
+  #
+  # Default: mokio.root_path ("/backend")
+  #
+  # Examples
+  # 1. Path:
+  # config.after_sign_in_path = {:user => {:path => "/path_for_all_users"}}
+  #
+  # 2. Method (add method to Mokio::ApplicationController):
+  # config.after_sign_in_path = {:user => {:method => :method_for_all_users}} => returns: controller.method_for_all_users
+  #
+  # Path based on user role examples:
+  #
+  # config.after_sign_in_path = {:user => {:roles => {:admin =>{:path => "/path_for_admin"}}}}
+  # config.after_sign_in_path = {:user => {:roles => {:admin =>{:method => :method_name_for_admin}}}} => returns: controller.method_name_for_admin
+
+  #config.after_sign_in_path = {:user => {}}
+>>>>>>> 22952500aa9e807b797a42cfbcf5988c1de15a22
 
 end