+++ lib/mokio/concerns/controllers/application.rb
@@ -27,20 +27,49 @@ module Mokio
             devise_parameter_sanitizer.for(:sign_up) {|u| u.permit(:email, :password, :password_confirmation, roles: [])}
           end
 
-        private 
           #
           # Standard devise after_sign_out_path_for
           #
           def after_sign_out_path_for(resource_or_scope) #:doc:
-            mokio.user_session_path
+            get_path_from_config(resource, Mokio.after_sign_out_path[resource_name], mokio.user_session_path)
           end
 
           #
           # Standard devise after_sign_in_path_for
           #
           def after_sign_in_path_for(resource) #:doc:
-            mokio.root_path
+            get_path_from_config(resource,  Mokio.after_sign_in_path[resource_name], mokio.root_path)
           end
+
+        private
+
+        def get_path_from_config(resource, config, default_path)
+          url = config.has_key?(:roles) ?
+            path_for_roles(config, resource)
+          :
+            get_path_from_hash(config, resource)
+
+          url ||=default_path
+          url
+        end
+
+        def path_for_roles(hash, resource)
+          url = nil
+          hash[:roles].each do |role_name, value|
+            if resource.roles.include? role_name
+              url = get_path_from_hash value, resource
+              break
+            end
+          end
+          url
+        end
+
+        def get_path_from_hash(hash, resource)
+          url = hash[:path].to_s if hash.has_key? :path
+          url = self.send(hash[:method]) if hash.has_key? :method
+          url
+        end
+
       end
     end
   end