+++ lib/mokio/concerns/controllers/base.rb
@@ -16,7 +16,18 @@ module Mokio
           attr_accessor :breadcrumbs_prefix
           attr_accessor :breadcrumbs_prefix_link
           before_filter :set_breadcrumbs_prefix
+
+
+          attr_accessor :external_parameters
+
         end
+
+
+
+        def external_parameters
+          [:dupa]
+        end
+
 
         #
         # Sending flash messages in X-Flash-Messages header.