+++ lib/mokio/concerns/models/base_article.rb
@@ -4,10 +4,15 @@ module Mokio
       #
       # Concern for Mokio::Article model
       #
-      module Article
+      module BaseArticle
         extend ActiveSupport::Concern
 
         included do
+          attr_accessor :form_active
+
+          def form_active
+            false
+          end
 
         end
       end