+++ lib/mokio/concerns/models/contact_template.rb
@@ -8,6 +8,10 @@ module Mokio
         extend ActiveSupport::Concern
 
         included do
+
+          belongs_to :base_contact
+          accepts_nested_attributes_for :base_contact
+
           belongs_to :contact
           accepts_nested_attributes_for :contact
         end